+++ sdk/spring/azure-spring-boot-starter-keyvault-secrets/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.7.0-beta.1 (Unreleased)
 ### New Features
-- Support unified properties as below when the associated "azure.keyvault.*" properties are not configured.([#22396](https://github.com/Azure/azure-sdk-for-java/issues/22396))
+- Support unified properties as below when the associated `azure.keyvault.*` properties are not configured.([#22396](https://github.com/Azure/azure-sdk-for-java/issues/22396))
     ```
     spring:
       cloud:
@@ -11,9 +11,13 @@
           client-secret:
           certificate-path:
           tenant-id:
+          authority-host:
+          environment:
     ```
 ### Breaking Changes
-- Deprecate property of "azure.keyvault.client-key", and provide ""azure.keyvault.client-secret" instead.
+- Deprecate property of `azure.keyvault.client-key`, and provide `azure.keyvault.client-secret` instead.
+- Property type of `spring.cloud.azure.environment` are changed from Enum to String. Supported values are "Azure", "AzureChina", "AzureGermany" and "AzureUSGovernment".
+- Support alternative configuration of `azure.keyvault.authority-host` from `spring.cloud.azure.authority-host` or `spring.cloud.azure.environment`.
 
 ## 3.6.0 (2021-06-23)
 ### Breaking Changes